Only generate with the -g flag
Cluster generation is a long process, and should never be
unexpected. With this change, clusters will only ever be
generated when passing the -g flag. This also overwrites
previously generated cluster data.

This commit also has the basic framework for loading or
generating clusters, though none of that is actually in yet.