Always predict with best chip
Previously, if there was a chip with doors, even if it's not the
best chip, we would use the doors chip. This caused the chip to
be rejected sometimes, when the best chip wouldn't've. The solution
is to always use the best chip, but if there's also a door chip,
combine the params from the best and door chips. It works well,
but makes parameter estimation very slow.